+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,8 @@
         android:icon="@mipmap/app_icon"
         android:theme="@style/AppTheme">
         <!--2016/3/31 churongShaw 静态注册蓝牙广播监听接收器-->
-        <receiver android:name="com.jcx.bcreceiver.BlueTReceiver" >
+        <receiver
+            android:name="com.jcx.bcreceiver.BlueTReceiver" >
             <intent-filter>
                 <action android:name="android.bluetooth.device.action.PAIRING_REQUEST" />
             </intent-filter>
@@ -38,15 +39,9 @@
         <uses-library android:name="android.test.runner"/>
         <!--2016/3/27-->
         <activity
-            android:name=".view.MainActivity"
-            android:label="@string/app_name"
-            >
+            android:name=".view.AllFilesActivity"
+            android:label="@string/app_name">
             <!--android:theme="@style/Theme.Appcompat" >-->
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name="com.zxing.activity.CaptureActivity"
@@ -55,9 +50,19 @@
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen"
             android:windowSoftInputMode="stateAlwaysHidden" >
         </activity>
-        <activity android:name=".view.CreateQRCodeActivity"
+        <activity
+            android:name=".view.CreateQRCodeActivity"
             android:theme="@style/AppTheme"
-            android:label="@string/activity_actionBar_label"></activity>
+            android:label="@string/activity_actionBar_label">
+        </activity>
+        <activity
+            android:name="com.jcx.view.MainActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
     </application>
 
     <!--2016/3/27 churongShaw 用于测试-->
