+++ app/src/main/res/menu/menu_main.xml
@@ -20,10 +20,10 @@
         <!--android:orderInCategory="100"-->
         <!--app:showAsAction="never"/>-->
     <!-- orderincategory属性的作用：每个item的优先级，值越大，优先级越低，不会放到overflow中 -->
-    <item android:id="@+id/downloading_file_manager"
-        android:title="@string/downloading_file_manager"
-        android:orderInCategory="100"
-        app:showAsAction="never"/>
+    <!--<item android:id="@+id/downloading_file_manager"-->
+        <!--android:title="@string/downloading_file_manager"-->
+        <!--android:orderInCategory="100"-->
+        <!--app:showAsAction="never"/>-->
     <item
         android:id="@+id/action_new_folder"
         android:title="@string/menu_new_folder"
